fix cases

+++ functions.c
@@ -47,7 +47,7 @@ int print_percent(void)
  */
 int print_d(va_list arg)
 {
-	int n = va_arg(arg, int);
+	long int n = va_arg(arg, int);
 	int divisor = 1, count = 0;
 
 	if (n == INT_MIN)
